added generic tutorial functionality, and refactored md tutorial to use it
