--- apps/ds-ui/src/components/common/md-tutorial/SpecialMarkdown.ts
@@ -1,10 +1,0 @@
-import { SectionDocumentation, PartType } from './TutorialTypes';
-
-export const specialMarkdownSection: SectionDocumentation = {
-  parts: [
-    {
-      type: PartType.text,
-      content: `All of the examples given so far are for "standard" markdown notation; anyone familiar with markdown would be familiar with the examples given, even if some (such as footnotes) might be less common than others (such as bold and italics). **Devouring Scripture** has also introduced some special notations that can be used specifically for the kind of text that often gets used in Christian contexts.`,
-    },
-  ],
-};
