--- apps/ds-ui/src/components/common/md-tutorial/TutorialTypes.tsx
@@ -1,11 +1,0 @@
-export enum PartType {
-  text = 'text',
-  example = 'example',
-}
-interface SectionPart {
-  type: PartType;
-  content: string;
-}
-export interface SectionDocumentation {
-  parts: SectionPart[];
-}
