+++ apps/ds-ui/src/components/common/tutorial/Tutorial.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { SectionDocumentation } from './TutorialTypes';
+import { RenderedSection } from './RenderedSection';
+
+interface ITutorial {
+  sections: SectionDocumentation[];
+}
+export const Tutorial = ({ sections }: ITutorial) => {
+  return (
+    <>
+      {sections.map((section) => (
+        <RenderedSection section={section} />
+      ))}
+    </>
+  );
+};
